# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ObfuscatorPresetLuaVersion = typeof ObfuscatorPresetLuaVersion[keyof typeof ObfuscatorPresetLuaVersion];


export const ObfuscatorPresetLuaVersion = {
  Lua51: 'Lua51',
  LuaU: 'LuaU',
} as const;

export interface ObfuscatorPreset {
  name: string;
  description: string;
  luaVersion: ObfuscatorPresetLuaVersion;
}

export type ObfuscateLuaInputPreset = typeof ObfuscateLuaInputPreset[keyof typeof ObfuscateLuaInputPreset];


export const ObfuscateLuaInputPreset = {
  Minify: 'Minify',
  Weak: 'Weak',
  Normal: 'Normal',
  Strong: 'Strong',
  Vmify: 'Vmify',
} as const;

export type ObfuscateLuaInputLuaVersion = typeof ObfuscateLuaInputLuaVersion[keyof typeof ObfuscateLuaInputLuaVersion];


export const ObfuscateLuaInputLuaVersion = {
  Lua51: 'Lua51',
  LuaU: 'LuaU',
} as const;

export interface ObfuscateLuaInput {
  /**
     * @minLength 1
     * @maxLength 262144
     */
  source: string;
  preset: ObfuscateLuaInputPreset;
  luaVersion?: ObfuscateLuaInputLuaVersion;
  prettyPrint?: boolean;
  /** @maxLength 120 */
  fileName?: string;
}

export interface ObfuscateLuaResult {
  source: string;
  preset: string;
  luaVersion: string;
  inputBytes: number;
  outputBytes: number;
  durationMs: number;
}

export interface ErrorResponse {
  error: string;
}