# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: obfuscator
    description: Lua obfuscation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /obfuscator/presets:
    get:
      operationId: listObfuscatorPresets
      tags: [obfuscator]
      summary: List obfuscation presets
      description: Returns the presets available in the bundled Prometheus engine
      responses:
        "200":
          description: Available presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ObfuscatorPreset"
  /obfuscator/obfuscate:
    post:
      operationId: obfuscateLua
      tags: [obfuscator]
      summary: Obfuscate Lua source
      description: Parses and transforms Lua source using the selected bundled preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObfuscateLuaInput"
      responses:
        "200":
          description: Obfuscated Lua source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObfuscateLuaResult"
        "400":
          description: Invalid input or Lua syntax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Source is too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Obfuscation engine error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ObfuscatorPreset:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        luaVersion:
          type: string
          enum: [Lua51, LuaU]
      required: [name, description, luaVersion]
    ObfuscateLuaInput:
      type: object
      properties:
        source:
          type: string
          minLength: 1
          maxLength: 262144
        preset:
          type: string
          enum: [Minify, Weak, Normal, Strong, Vmify]
        luaVersion:
          type: string
          enum: [Lua51, LuaU]
        prettyPrint:
          type: boolean
        fileName:
          type: string
          maxLength: 120
      required: [source, preset]
    ObfuscateLuaResult:
      type: object
      properties:
        source:
          type: string
        preset:
          type: string
        luaVersion:
          type: string
        inputBytes:
          type: number
        outputBytes:
          type: number
        durationMs:
          type: number
      required:
        [source, preset, luaVersion, inputBytes, outputBytes, durationMs]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]