# lib/api-zod/src/generated/types/obfuscateLuaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObfuscateLuaInputLuaVersion } from './obfuscateLuaInputLuaVersion';
import type { ObfuscateLuaInputPreset } from './obfuscateLuaInputPreset';

export interface ObfuscateLuaInput {
  /**
     * @minLength 1
     * @maxLength 262144
     */
  source: string;
  preset: ObfuscateLuaInputPreset;
  luaVersion?: ObfuscateLuaInputLuaVersion;
  prettyPrint?: boolean;
  /** @maxLength 120 */
  fileName?: string;
}