# lib/api-zod/src/generated/types/obfuscateLuaResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ObfuscateLuaResult {
  source: string;
  preset: string;
  luaVersion: string;
  inputBytes: number;
  outputBytes: number;
  durationMs: number;
}