# lib/api-zod/src/generated/types/obfuscatorPreset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObfuscatorPresetLuaVersion } from './obfuscatorPresetLuaVersion';

export interface ObfuscatorPreset {
  name: string;
  description: string;
  luaVersion: ObfuscatorPresetLuaVersion;
}